Handle html links and inline styles
